Added script part for loading page
When we show the demo on exhibitions, we found one problem. After user clicks the "Match Me" button, the webpage stays same and user does not have an idea of what's happening. So we made this modification. When user clicks the "Match Me" button, it will show a page saying "Your Celebrity Match is extracting your tweets or instagram images you posted to predict your personality........." along with explanation of how things work. In this way, if user entered an account that is not cached in db thus need to wait for long time, the user can read the explanations to pass time instead of just waiting at the index page.